Move set/term/scored out of input
