+++ app/public/playbooksDemo.ts
@@ -2,9 +2,9 @@ import { MetgeneMetaboliteTable } from '@/components/MW/metgene_metabolite_table
 import { MetGeneRxnTable } from '@/components/MW/metgene_rxn_table'
 import { MetGeneStudyTable } from '@/components/MW/metgene_study_table'
 import { MetGeneSummary } from '@/components/MW/metgene_summary'
-import { ScoredDrugs, ScoredGenes, ScoredTissues } from '@/components/core/input/scored'
-import { GeneSet, RegulatoryElementSet, VariantSet } from '@/components/core/input/set'
-import { DrugTerm, GeneTerm, PhenotypeTerm, VariantTerm } from '@/components/core/input/term'
+import { ScoredDrugs, ScoredGenes, ScoredTissues } from '@/components/core/scored'
+import { GeneSet, RegulatoryElementSet, VariantSet } from '@/components/core/set'
+import { DrugTerm, GeneTerm, PhenotypeTerm, VariantTerm } from '@/components/core/term'
 import { GeneCountMatrix } from '@/components/data/gene_count_matrix'
 import { GeneSignature } from '@/components/data/gene_signature'
 import { GlyGenProteinResponseNode } from '@/components/gly_gen'
