+++ components/MW/PPI_StringDB/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { MetaNode } from '@/spec/metanode'
-import { GeneTerm } from '@/components/core/input/term'
-import { GeneSet } from '@/components/core/input/set'
+import { GeneTerm } from '@/components/core/term'
+import { GeneSet } from '@/components/core/set'
 import { GraphPlot } from '@/components/viz/graph'
 import { metabolomicsworkbench_icon } from '@/icons'
 import { z } from 'zod'
