+++ components/MW/geneidconv/index.tsx
@@ -1,6 +1,6 @@
 import { MetaNode } from '@/spec/metanode'
-import { GeneTerm } from '@/components/core/input/term'
-import { GeneSet } from '@/components/core/input/set'
+import { GeneTerm } from '@/components/core/term'
+import { GeneSet } from '@/components/core/set'
 import { ConvertedGeneID } from '@/components/MW/ConvertedGeneID'
 import { uniqJsonSubset } from '@/components/MW/ConvertedGeneID'
 import { metabolomicsworkbench_icon } from '@/icons'
