+++ components/MW/metgene_metabolites/index.tsx
@@ -1,7 +1,7 @@
 import { MetaNode } from '@/spec/metanode'
-import { GeneSet } from '@/components/core/input/set'
+import { GeneSet } from '@/components/core/set'
 import { MetgeneMetaboliteTable } from '../metgene_metabolite_table'
-import { MetaboliteSet } from '@/components/core/input/set'
+import { MetaboliteSet } from '@/components/core/set'
 import { metgene_icon } from '@/icons'
 import { MetGeneSummary } from '@/components/MW/metgene_summary'
 import * as array from '@/utils/array'
