+++ components/core/scored/index.tsx
