+++ components/core/scored/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "term",
+  "name": "@playbook-partnership/scored",
   "version": "1.0.0",
   "license": "CC-BY-NC-SA-4.0",
   "author": "Daniel J. B. Clarke <danieljbclarkemssm@gmail.com>",
