+++ components/core/term/index.tsx
