+++ components/data/drug_matrix_transpose/index.tsx
@@ -4,7 +4,7 @@ import { z } from 'zod'
 import * as dict from '@/utils/dict'
 import * as array from '@/utils/array'
 import { Table, Cell, Column, EditableCell } from '@/app/components/Table'
-import { DrugSet } from '@/components/core/input/set'
+import { DrugSet } from '@/components/core/set'
 import { FileC, FilePrompt, FileURL } from '@/components/core/file'
 import { downloadBlob } from '@/utils/download'
 import { file_icon, file_transfer_icon, gmt_icon } from '@/icons'
