+++ components/data/gene_signature/index.tsx
@@ -7,8 +7,8 @@ import Matrix from '@/app/components/Matrix'
 import { datafile_icon, filter_icon, differential_expression_icon, file_transfer_icon, file_icon } from '@/icons'
 import { downloadUrl } from '@/utils/download'
 import { GMT } from '../gene_matrix_transpose'
-import { GeneSet } from '@/components/core/input/set'
-import { ScoredGenes } from '@/components/core/input/scored'
+import { GeneSet } from '@/components/core/set'
+import { ScoredGenes } from '@/components/core/scored'
 import SafeRender from '@/utils/saferender'
 import { clientLoadExample } from  '@/components/data/gene_signature/api/GTEx_aging_signature_limma.tsv/client'
