+++ components/filters/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { z } from 'zod'
 import { MetaNode } from '@/spec/metanode'
-import { DrugSet, GeneSet, GlycanSet, ProteinSet, VariantSet, RegulatoryElementSet } from '@/components/core/input/set'
+import { DrugSet, GeneSet, GlycanSet, ProteinSet, VariantSet, RegulatoryElementSet } from '@/components/core/set'
 import { Disease, Drug, Gene, Glycan, Pathway, Phenotype, Protein, Tissue, Variant, RegulatoryElement } from '@/components/core/primitives'
-import { DiseaseTerm, DrugTerm, GeneTerm, GlycanTerm, PathwayTerm, PhenotypeTerm, ProteinTerm, TissueTerm, VariantTerm, RegulatoryElementTerm } from '@/components/core/input/term'
-import { ScoredDrugs, ScoredGenes, ScoredGlycans, ScoredDiseases, ScoredPathways, ScoredPhenotypes, ScoredProteins, ScoredTissues, ScoredVariants, ScoredRegulatoryElement } from '@/components/core/input/scored'
+import { DiseaseTerm, DrugTerm, GeneTerm, GlycanTerm, PathwayTerm, PhenotypeTerm, ProteinTerm, TissueTerm, VariantTerm, RegulatoryElementTerm } from '@/components/core/term'
+import { ScoredDrugs, ScoredGenes, ScoredGlycans, ScoredDiseases, ScoredPathways, ScoredPhenotypes, ScoredProteins, ScoredTissues, ScoredVariants, ScoredRegulatoryElement } from '@/components/core/scored'
 import { Table, Cell, Column } from '@/app/components/Table'
 import * as dict from '@/utils/dict'
 import * as math from '@/utils/math'
