+++ components/service/archs4/index.tsx
@@ -1,7 +1,7 @@
 import { MetaNode } from '@/spec/metanode'
-import { ScoredGenes, ScoredTissues } from '@/components/core/input/scored'
+import { ScoredGenes, ScoredTissues } from '@/components/core/scored'
 import { archs4_icon, weighted_icon } from '@/icons'
-import { GeneTerm } from '@/components/core/input/term'
+import { GeneTerm } from '@/components/core/term'
 import read_csv from '@/utils/csv'
 import { z } from 'zod'
 import { Cell, Column, Table } from '@/app/components/Table'
