+++ components/service/kf/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { MetaNode } from '@/spec/metanode'
 import { z } from 'zod'
 import python from '@/utils/python'
-import { GeneTerm } from '@/components/core/input/term'
+import { GeneTerm } from '@/components/core/term'
 import { GeneInfo, GeneInfoFromGeneTerm } from '@/components/service/mygeneinfo'
 import { Table, Cell, Column } from '@/app/components/Table'
 import { tumor_icon } from '@/icons'
