+++ components/service/regulatoryElementInfo/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { MetaNode } from '@/spec/metanode'
-import { RegulatoryElementTerm } from '@/components/core/input/term'
-import { GeneSet, RegulatoryElementSet, VariantSet } from '@/components/core/input/set'
+import { RegulatoryElementTerm } from '@/components/core/term'
+import { GeneSet, RegulatoryElementSet, VariantSet } from '@/components/core/set'
 import { z } from 'zod'
 import { linkeddatahub_icon, datafile_icon } from '@/icons'
 import { Table, Cell, Column} from '@/app/components/Table'
