+++ components/service/sigcom-lincs/index.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { MetaNode } from '@/spec/metanode'
 import { z } from 'zod'
 import * as dict from '@/utils/dict'
-import { ScoredDrugs, ScoredGenes } from '@/components/core/input/scored'
-import { GeneSet } from '@/components/core/input/set'
+import { ScoredDrugs, ScoredGenes } from '@/components/core/scored'
+import { GeneSet } from '@/components/core/set'
 import { drug_icon, gene_icon, lincs_icon, set_icon, weighted_icon } from '@/icons'
 
 const sigcom_lincs_url = 'https://maayanlab.cloud/sigcom-lincs'
