+++ components/service/variantinfo/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { MetaNode } from '@/spec/metanode'
-import { VariantTerm } from '@/components/core/input/term'
-import { RegulatoryElementTerm } from '@/components/core/input/term'
-import { VariantSet } from '@/components/core/input/set'
+import { VariantTerm } from '@/components/core/term'
+import { RegulatoryElementTerm } from '@/components/core/term'
+import { VariantSet } from '@/components/core/set'
 import { Table, Cell, Column} from '@/app/components/Table'
 import { z } from 'zod'
 import { linkeddatahub_icon } from '@/icons'
