+++ docs/component_templates/myprompt/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { MetaNode } from '@/spec/metanode'
-import { GeneTerm } from '@/components/core/input/term'
+import { GeneTerm } from '@/components/core/term'
 
 // A unique name for your prompt is used here
 export const MyPrompt = MetaNode('MyPrompt')
