+++ docs/component_templates/mypythonidentity/index.tsx
@@ -1,6 +1,6 @@
 import python from '@/utils/python'
 import { MetaNode } from '@/spec/metanode'
-import { GeneTerm } from '@/components/core/input/term'
+import { GeneTerm } from '@/components/core/term'
 
 // A unique name for your resolver is used here
 export const MyPythonIdentity = MetaNode('MyPythonIdentity')
